BFDDW-9 fix: resolve PostgreSQL issue

+++ src/main/java/com/bloomberg/fxdeals/dto/FxDealRequestDto.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record FxDealRequestDto(
@@ -24,5 +25,5 @@ public record FxDealRequestDto(
 
         @NotNull(message = "Amount is required")
         @Positive(message = "Amount must be positive")
-        Double amount
+        BigDecimal amount
 ) {}
